pyproject.toml
[project]
name = "blueos-foxglove"
version = "0.1.0"
description = "MAVLink to Foxglove bridge"
requires-python = ">=3.9"
dependencies = [
    "foxglove-websocket>=0.1.0",
    "foxglove-sdk>=0.7.1",
    "eclipse-zenoh==1.3.4",
    "asyncio>=3.4.3",
    "aiohttp>=3.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["."]